requirements.txt
Django==2.1.7
django-cors-headers==2.5.2
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
Pillow==5.4.1
pytz==2018.9
